+++ asf_search/ASFSearchOptions/validator_map.py
@@ -79,6 +79,8 @@ validator_map = {
     'shortName': parse_string_list,
     'dataset': parse_string_list,
     'cmr_keywords': parse_cmr_keywords_list,
+    'ariaVersion': parse_string,
+    'tileID': parse_string_list,
     # S1 Inteferrogram Filters
     'temporalBaselineDays': parse_string_list,
     # Opera Burst Filters