+++ asf_search/CMR/field_map.py
@@ -38,6 +38,7 @@ field_map = {
     'collections':                  {'key': 'echo_collection_id[]',     'fmt': '{0}'},
     'shortName':                    {'key': 'shortName',                'fmt': '{0}'},
     'temporalBaselineDays':         {'key': 'attribute[]',              'fmt': 'int,TEMPORAL_BASELINE_DAYS,{0}'},  # noqa F401
+    'ariaVersion':                  {'key': 'attribute[]',              'fmt': 'string,VERSION,{0}'},
     # SLC BURST fields
     'absoluteBurstID':              {'key': 'attribute[]',              'fmt': 'int,BURST_ID_ABSOLUTE,{0}'},
     'relativeBurstID':              {'key': 'attribute[]',              'fmt': 'int,BURST_ID_RELATIVE,{0}'},
@@ -51,4 +52,5 @@ field_map = {
     'jointObservation':             {'key': 'attribute[]',             'fmt': 'string,JOINT_OBSERVATION,{0}'},
     'rangeBandwidth':               {'key': 'attribute[]',             'fmt': 'string,RANGE_BANDWIDTH_CONCAT,{0}'},
     'productionConfiguration':      {'key': 'attribute[]',             'fmt': 'string,PRODUCTION_PIPELINE,{0}'},
+    'tileID':                       {'key': 'attribute[]',             'fmt': 'string,MGRS_TILE_ID,{0}'},
 }