+++ asf_search/search/search.py
@@ -46,6 +46,7 @@ def search(
     ] = None,
     season: Tuple[int, int] = None,
     start: Union[datetime.datetime, str] = None,
+    ariaVersion: str = None,
     absoluteBurstID: Union[int, Sequence[int]] = None,
     relativeBurstID: Union[int, Sequence[int]] = None,
     fullBurstID: Union[str, Sequence[str]] = None,
@@ -56,12 +57,13 @@ def search(
     sideBandPolarization: Union[str, Sequence[str]] = None,
     rangeBandwidth: Union[str, Sequence[str]] = None,
     jointObservation: bool = None,
-    productionConfiguration: Union[Literal["PR", "UR"], Sequence[Literal["PR", "UR"]]] = None,
+    productionConfiguration: Union[Literal['PR', 'UR'], Sequence[Literal['PR', 'UR']]] = None,
     dataset: Union[str, Sequence[str]] = None,
     collections: Union[str, Sequence[str]] = None,
     shortName: Union[str, Sequence[str]] = None,
     cmr_keywords: Union[Tuple[str, str], Sequence[Tuple[str, str]]] = None,
     maxResults: int = None,
+    tileID: Union[str, Sequence[str]] = None,
     opts: ASFSearchOptions = None,
 ) -> ASFSearchResults:
     """
@@ -79,6 +81,8 @@ def search(
         For ALOS, ERS-1, ERS-2, JERS-1, and RADARSAT-1, Sentinel-1A, Sentinel-1B
         this value corresponds to the orbit count within the orbit cycle.
         For UAVSAR it is the Flight ID.
+    ariaVersion:
+        For ARIAS1GUNW this value describes the version a product was generated under
     asfFrame:
         This is primarily an ASF / JAXA frame reference. However,
         some platforms use other conventions. See ‘frame’ for ESA-centric frame searches.
@@ -114,6 +118,7 @@ def search(
     granule_list:
         List of specific granules.
         Search results may include several products per granule name.
+        Supports wildcard queries (*/?)
     groupID:
         Identifier used to find products considered to
         be of the same scene but having different granule names
@@ -158,6 +163,8 @@ def search(
     temporalBaselineDays:
         List of temporal baselines,
         used for Sentinel-1 Interferogram (BETA)
+    tileID:
+        For DIST-ALERT-S1 product type products
     maxResults:
         The maximum number of results to be returned by the search
     opts:
@@ -185,7 +192,6 @@ def search(
         results.searchOptions = page.searchOptions
         perf = time.time()
 
-
     if not results.searchComplete:
         msg = (
             'Results may be incomplete due to a search error. '
@@ -193,7 +199,6 @@ def search(
         )
 
         ASF_LOGGER.error(msg)
-        
 
     try:
         results.sort(key=lambda p: p.get_sort_keys(), reverse=True)